+++ crates/ffi/src/platforms/windows/AdGuard.FilterListManagerProtobuf/AdGuard.FilterListManagerProtobuf/RustInterface/FFIMethod.cs
@@ -14,6 +14,7 @@
         UpdateFilters,
         ForceUpdateFiltersByIds,
         FetchFilterListMetadata,
+        FetchFilterListMetadataWithBody,
         LiftUpDatabase,
         GetAllTags,
         GetAllGroups,