+++ crates/filter-list-manager/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.1] - 2025-02-24
+
+### Added
+- `fetch_filter_list_metadata_with_body` method for fetch filter metadata with filter body
+
+[1.4.1]: https://github.com/AdguardTeam/FilterListManager/compare/flm-1.3.1...flm-1.4.1
+
 ## [1.3.1] - 2025-01-31
 
 ### Added